Separate specs component
